+++ CHANGELOG.md
@@ -1,10 +1,18 @@
 # Changelog
 
+## [v1.2.0] - 2021-11-08
+### Added
+- Options to set `Assign` default anchor logic
+- `Assign` Nearest Heading logic
+
+### Changed
+- `Quick Goto - Assign` confirmation dialog with anchor logic options
+
 ## [v1.1.0] - 2021-10-29
 
 ### Added
-- Options to enable silence `Assign`
-- Options to set `Assign` hotkeys
+- Options to enable `Assign` silently
+- Options to set `Assign` keyboard shortcuts
 - Support Go To a specific header of the note (no GUI)
 - Support Go To a specific header of the same not
 
@@ -23,3 +31,7 @@
 ### Added
 - Menu Item and Shortcut for `Quick Goto`
 - Menu Item for `Quick Goto - Assign`
+
+[v1.2.0]: https://github.com/hieuthi/joplin-plugin-quick-goto/compare/v1.1.0...v1.2.0
+[v1.1.0]: https://github.com/hieuthi/joplin-plugin-quick-goto/compare/v1.0.1...v1.1.0
+[v1.0.1]: https://github.com/hieuthi/joplin-plugin-quick-goto/releases/tag/v1.0.1